pyproject.toml
[project]
name = "madeinabyss-minecraft-search"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.40.4",
    "faiss-cpu>=1.11.0.post1",
    "langchain>=0.3.27",
    "langchain-aws>=0.2.30",
    "mcpi>=1.2.1",
    "mcrcon>=0.7.0",
    "nbtlib>=2.0.4",
    "numpy>=2.3.2",
    "tqdm>=4.67.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mvs"]

[tool.ruff.lint]
unfixable = [
    "F401", # unused import
    "F841", # unused variable
]
